Break logic into sperate classes
